requirements.txt
streamlit==1.24.0
langchain-core==0.3.46
langchain-groq==0.2.5
langchain-openai==0.3.9
langchain-community==0.3.20
openai==1.66.3
chromadb==0.4.15
PyPDF2==3.0.1
python-dotenv==1.0.0
beautifulsoup4==4.12.3
youtube-transcript-api==0.6.2
yt-dlp==2024.3.10
pypdf
langchain-chroma==0.2.2
